backend/python/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
httpx
requests
openai
beautifulsoup4
pgvector
python-multipart
